Moving Constants into folder
