+++ webAO/constants/characters.js
